4.015 (3/6/2011)
Lobby (360): Added Console_GetRoomDetails_360
Lobby (360): Added RoomMemberInfo::roomMemberName
Lobby (PS3): Exposed supportHostMigration parameter when creating rooms.
Autopatcher: Add CacheMostRecentPatch() to AutopatcherServer. This keeps the most recent patch operation in memory, as well as the start and end date the most recent patch occurred. When the lastUpdateDate is passed to AutopatcherClient::PatchApplication() and the corresponding patch is the most recent, disk access is not needed on the server, greatly improving performance.